+++ template/{{cookiecutter.year}}/src/day10.rs
@@ -1,8 +1,8 @@
-pub fn day10p1(_input: &str) -> usize {
+pub fn day10p1(_input: &str) -> u32 {
     0
 }
 
-pub fn day10p2(_input: &str) -> usize {
+pub fn day10p2(_input: &str) -> u32 {
     0
 }
