+++ template/{{cookiecutter.year}}/src/day18.rs
@@ -1,8 +1,8 @@
-pub fn day18p1(_input: &str) -> usize {
+pub fn day18p1(_input: &str) -> u32 {
     0
 }
 
-pub fn day18p2(_input: &str) -> usize {
+pub fn day18p2(_input: &str) -> u32 {
     0
 }
