+++ template/{{cookiecutter.year}}/src/day25.rs
@@ -1,8 +1,8 @@
-pub fn day25p1(_input: &str) -> usize {
+pub fn day25p1(_input: &str) -> u32 {
     0
 }
 
-pub fn day25p2(_input: &str) -> usize {
+pub fn day25p2(_input: &str) -> u32 {
     0
 }
